+++ packages/react/src/components/Toolbar/index.tsx
@@ -1095,6 +1095,7 @@ const Toolbar: React.FC<{
       if (name === "splitColumn") {
         return (
           <Button
+            id="splitColumn"
             iconId={name}
             tooltip={tooltip}
             key={name}
@@ -1124,6 +1125,7 @@ const Toolbar: React.FC<{
           <>
             <DataVerificationPortal visible={showDataValidation} />
             <Button
+              id="dataVerification"
               iconId={name}
               tooltip={tooltip}
               key={name}
@@ -1780,7 +1782,8 @@ const Toolbar: React.FC<{
                           refs.cellInput.current!,
                           d,
                           "tb",
-                          value
+                          value,
+                          refs.canvas.current!.getContext("2d")!
                         );
                       });
                       setOpen(false);