+++ packages/react/CHANGELOG.md
@@ -1,5 +1,45 @@
 # @fileverse-dev/fortune-react
 
+## 1.2.27
+
+### Patch Changes
+
+- Tab select fx editor
+- Updated dependencies
+  - @fileverse-dev/fortune-core@1.2.27
+
+## 1.2.26
+
+### Patch Changes
+
+- N/A
+- Updated dependencies
+  - @fileverse-dev/fortune-core@1.2.26
+
+## 1.2.25
+
+### Patch Changes
+
+- move cell disabled for flv
+- Updated dependencies
+  - @fileverse-dev/fortune-core@1.2.25
+
+## 1.2.24
+
+### Patch Changes
+
+- N/A
+- Updated dependencies
+  - @fileverse-dev/fortune-core@1.2.24
+
+## 1.2.23
+
+### Patch Changes
+
+- Fx header formula suggestion
+- Updated dependencies
+  - @fileverse-dev/fortune-core@1.2.23
+
 ## 1.2.22
 
 ### Patch Changes